requirements.txt
# Reddit User Persona Generator Dependencies
praw==7.8.1
nltk==3.8.1
spacy==3.7.2
pandas==2.1.1
argparse==1.4.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
tqdm==4.66.1
colorama==0.4.6